+++ docs/manifest.md
